improve file download verbositiy and file replace mimeType check. Only the same files can be replaced with each other

+++ modules/admin/src/apis/StorageController.php
@@ -15,6 +15,7 @@ use yii\caching\DbDependency;
 use luya\admin\filters\TinyCrop;
 use luya\admin\filters\MediumThumbnail;
 use luya\helpers\FileHelper;
+use yii\web\BadRequestHttpException;
 
 /**
  * Filemanager and Storage API.
@@ -198,6 +199,14 @@ class StorageController extends RestController
         if ($file = Yii::$app->storage->getFile($fileId)) {
             $serverSource = $file->getServerSource();
             if (is_uploaded_file($raw['tmp_name'])) {
+            	
+            	// check for same extension / mimeType
+            	$fileData = Yii::$app->storage->ensureFileUpload($raw['tmp_name'], $raw['name']);
+            	
+            	if ($fileData['mimeType'] != $file->mimeType) {
+            		throw new BadRequestHttpException("The type must be the same as the original file in order to replace.");
+            	}
+            	
                 if (Storage::replaceFile($serverSource, $raw['tmp_name'], $raw['name'])) {
                     foreach (Yii::$app->storage->findImages(['file_id' => $file->id]) as $img) {
                         Storage::removeImage($img->id, false);
